Relax assert on loading of checkpoint variables for ic3 module

+++ egenerator/ic3/reconstruction.py
@@ -352,7 +352,7 @@ class EventGeneratorReconstruction(icetray.I3ConditionalModule):
                     "config": {
                         "sigmas": [],
                         "uniform_parameters": self.parameter_boundaries,
-                        "float_precision": "float64",
+                        "float_precision": "float32",
                     },
                 }
             ]
